# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenAgents API",
        "description": "The OpenAgents API",
        "contact": {
            "name": "OpenAgents",
            "url": "https://x.com/OpenAgentsInc"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://openagents.com/api/v1",
            "description": "OpenAgents API server"
        }
    ],
    "paths": {
        "/agents": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "List agents",
                "description": "Returns a list of agents owned by the user.",
                "operationId": "listAgents",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Agent"
                                            }
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agent"
                ],
                "summary": "Create agent",
                "operationId": "createAgent",
                "requestBody": {
                    "description": "Agent information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "instructions"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "instructions": {
                                        "type": "string"
                                    },
                                    "welcome_message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "agent_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/agents/{id}": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Retrieve agent",
                "description": "Retrieves an agent.",
                "operationId": "getAgentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of agent to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Agent"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agent"
                ],
                "summary": "Modify agent",
                "operationId": "updateAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of agent that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Agent data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "instructions": {
                                        "type": "string"
                                    },
                                    "welcome_message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Agent"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agent"
                ],
                "summary": "Delete agent",
                "operationId": "deleteAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agent id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Agent": {
                "properties": {
                    "id": {
                        "description": "Agent ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Agent name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Agent description",
                        "type": "string"
                    },
                    "instructions": {
                        "description": "Agent instructions",
                        "type": "string"
                    },
                    "welcome_message": {
                        "description": "Agent welcome message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}